fix conflicts

+++ packages/trpc/server/routers/viewer.tsx
@@ -1,9 +1,12 @@
 import { AppCategories, BookingStatus, MembershipRole, Prisma } from "@prisma/client";
+import { IdentityProvider, UserPermissionRole } from "@prisma/client";
 import _ from "lodash";
+import { authenticator } from "otplib";
 import { JSONObject } from "superjson/dist/types";
 import { z } from "zod";
 
 import app_RoutingForms from "@calcom/app-store/ee/routing_forms/trpc-router";
+import { deleteStripeCustomer } from "@calcom/app-store/stripepayment/lib/customer";
 import stripe, { closePayments } from "@calcom/app-store/stripepayment/lib/server";
 import getApps, { getLocationOptions } from "@calcom/app-store/utils";
 import { cancelScheduledJobs } from "@calcom/app-store/zapier/lib/nodeScheduler";
@@ -11,7 +14,6 @@ import { getCalendarCredentials, getConnectedCalendars } from "@calcom/core/Cale
 import dayjs from "@calcom/dayjs";
 import { sendCancelledEmails, sendFeedbackEmail } from "@calcom/emails";
 import { isPrismaObjOrUndefined, parseRecurringEvent } from "@calcom/lib";
-import { CAL_URL } from "@calcom/lib/constants";
 import hasKeyInMetadata from "@calcom/lib/hasKeyInMetadata";
 import jackson from "@calcom/lib/jackson";
 import {
@@ -29,6 +31,8 @@ import { isTeamOwner } from "@calcom/lib/server/queries/teams";
 import slugify from "@calcom/lib/slugify";
 import prisma, { baseEventTypeSelect, bookingMinimalSelect } from "@calcom/prisma";
 import { resizeBase64Image } from "@calcom/web/server/lib/resizeBase64Image";
+
+import { ErrorCode, verifyPassword } from "@lib/auth";
 
 import { TRPCError } from "@trpc/server";
 
@@ -103,15 +107,75 @@ const loggedInViewerRouter = createProtectedRouter()
     },
   })
   .mutation("deleteMe", {
-    async resolve({ ctx }) {
-      // Remove me from Stripe
-
-      // Remove my account
-      await ctx.prisma.user.delete({
+    input: z.object({
+      password: z.string(),
+      totpCode: z.string().optional(),
+    }),
+    async resolve({ input, ctx }) {
+      // Check if input.password is correct
+      const user = await prisma.user.findUnique({
         where: {
-          id: ctx.user.id,
+          email: ctx.user.email.toLowerCase(),
         },
       });
+
+      if (!user) {
+        throw new Error(ErrorCode.UserNotFound);
+      }
+
+      if (user.identityProvider !== IdentityProvider.CAL) {
+        throw new Error(ErrorCode.ThirdPartyIdentityProviderEnabled);
+      }
+
+      if (!user.password) {
+        throw new Error(ErrorCode.UserMissingPassword);
+      }
+
+      const isCorrectPassword = await verifyPassword(input.password, user.password);
+      if (!isCorrectPassword) {
+        throw new Error(ErrorCode.IncorrectPassword);
+      }
+
+      if (user.twoFactorEnabled) {
+        if (!input.totpCode) {
+          throw new Error(ErrorCode.SecondFactorRequired);
+        }
+
+        if (!user.twoFactorSecret) {
+          console.error(`Two factor is enabled for user ${user.id} but they have no secret`);
+          throw new Error(ErrorCode.InternalServerError);
+        }
+
+        if (!process.env.CALENDSO_ENCRYPTION_KEY) {
+          console.error(`"Missing encryption key; cannot proceed with two factor login."`);
+          throw new Error(ErrorCode.InternalServerError);
+        }
+
+        const secret = symmetricDecrypt(user.twoFactorSecret, process.env.CALENDSO_ENCRYPTION_KEY);
+        if (secret.length !== 32) {
+          console.error(
+            `Two factor secret decryption failed. Expected key with length 32 but got ${secret.length}`
+          );
+          throw new Error(ErrorCode.InternalServerError);
+        }
+
+        const isValidToken = authenticator.check(input.totpCode, secret);
+        if (!isValidToken) {
+          throw new Error(ErrorCode.IncorrectTwoFactorCode);
+        }
+        // If user has 2fa enabled, check if input.totpCode is correct
+        // If it is, delete the user from stripe and database
+
+        // Remove me from Stripe
+        await deleteStripeCustomer(user).catch(console.warn);
+
+        // Remove my account
+        await ctx.prisma.user.delete({
+          where: {
+            id: ctx.user.id,
+          },
+        });
+      }
       return;
     },
   })
